+++ VpnHood.Server/SslCertificateManager.cs
@@ -3,7 +3,7 @@ using System.Collections.Concurrent;
 using System.Net;
 using System.Security.Cryptography.X509Certificates;
 using System.Threading.Tasks;
-using VpnHood.Server.Exceptions;
+using VpnHood.Common.Exceptions;
 
 namespace VpnHood.Server
 {
@@ -12,10 +12,12 @@ namespace VpnHood.Server
         private readonly IAccessServer _accessServer;
         private readonly ConcurrentDictionary<IPEndPoint, X509Certificate2> _certificates = new();
         private readonly Lazy<X509Certificate2> _maintenanceCertificate = new(InitMaintenanceCertificate);
+        private readonly TimeSpan _maintenanceCheckInterval;
         private DateTime _lastMaintenanceTime = DateTime.Now;
 
-        public SslCertificateManager(IAccessServer accessServer)
+        public SslCertificateManager(IAccessServer accessServer, TimeSpan maintenanceCheckInterval)
         {
+            _maintenanceCheckInterval = maintenanceCheckInterval;
             _accessServer = accessServer;
         }
 
@@ -32,10 +34,10 @@ namespace VpnHood.Server
         public async Task<X509Certificate2> GetCertificate(IPEndPoint ipEndPoint)
         {
             // check maintenance mode
-            if (_accessServer.IsMaintenanceMode && (DateTime.Now - _lastMaintenanceTime).TotalMinutes < 1)
+            if (_accessServer.IsMaintenanceMode && (DateTime.Now - _lastMaintenanceTime) < _maintenanceCheckInterval)
                 return _maintenanceCertificate.Value;
 
-            // find in cache 
+            // find in cache
             if (_certificates.TryGetValue(ipEndPoint, out var certificate))
                 return certificate;
 
@@ -44,7 +46,8 @@ namespace VpnHood.Server
             {
                 var certificateData = await _accessServer.GetSslCertificateData(ipEndPoint);
                 certificate = new X509Certificate2(certificateData);
-                _certificates.TryAdd(ipEndPoint, certificate);
+                if (_maintenanceCheckInterval != TimeSpan.Zero) // test mode should not use cache
+                    _certificates.TryAdd(ipEndPoint, certificate);
                 return certificate;
             }
             catch (MaintenanceException)