+++ CHANGELOG.md
@@ -1,4 +1,21 @@
-﻿# v2.1.276
+﻿# v2.2.283
+### Client
+* Feature: Allow to have multi-endpoints in AccessToken
+* Feature: Create IPv6 tunnel when a client has access to a server by IPv6
+* Feature: Add "Exclude Local Network" to UI settings
+* Fix: UDP Channel
+
+### Server
+* Feature: Dynamic configuration from AccessServer
+* Feature: Multi listeners for different EndPoints
+* Fix: Few bug in disposing
+* Fix: linux: systemctl restart VpnHoodServer 
+
+# v2.1.276
+* Feature: IPv6 Support
+* Fix: Some packet loss in ping 
+
+# v2.1.276
 * Feature: IPv6 Support
 * Fix: Some packet loss in ping 
 
@@ -269,6 +286,7 @@
 
 # v1.1.75
 * Initial Release
+